Cleaned up code. Fixed scroll view for gallery. Organized scripts in directory structure.
