+++ U3d_Image360Viewer/Assets/PhotoViewer/Scripts/Extensions.cs
@@ -1,32 +1,22 @@
-﻿using System.Runtime.CompilerServices;
-using System.Threading.Tasks;
-using UnityEngine;
+﻿using UnityEngine;
 
-public static class Extensions
+namespace PhotoViewer.Scripts
 {
-    public static Texture2D TextureFromSprite(this Sprite sprite)
+    public static class Extensions
     {
-        if(sprite.rect.width != sprite.texture.width){
-            Texture2D newText = new Texture2D((int)sprite.rect.width,(int)sprite.rect.height);
-            Color[] newColors = sprite.texture.GetPixels((int)sprite.textureRect.x, 
-                (int)sprite.textureRect.y, 
-                (int)sprite.textureRect.width, 
-                (int)sprite.textureRect.height );
-            newText.SetPixels(newColors);
-            newText.Apply();
-            return newText;
-        } else
-            return sprite.texture;
+        public static Texture2D TextureFromSprite(this Sprite sprite)
+        {
+            if(sprite.rect.width != sprite.texture.width){
+                Texture2D newText = new Texture2D((int)sprite.rect.width,(int)sprite.rect.height);
+                Color[] newColors = sprite.texture.GetPixels((int)sprite.textureRect.x, 
+                    (int)sprite.textureRect.y, 
+                    (int)sprite.textureRect.width, 
+                    (int)sprite.textureRect.height );
+                newText.SetPixels(newColors);
+                newText.Apply();
+                return newText;
+            } else
+                return sprite.texture;
+        }
     }
-    
-    public static TaskAwaiter GetAwaiter(this AsyncOperation asyncOp)
-    {
-        var tcs = new TaskCompletionSource<object>();
-        asyncOp.completed += obj => { tcs.SetResult(null); };
-        return ((Task) tcs.Task).GetAwaiter();
-    }
-    
-    // var getRequest = UnityWebRequest.Get("http://www.google.com");
-    // await getRequest.SendWebRequest();
-    // var result = getRequest.downloadHandler.text;
 }
